backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-ollama
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters
python-dotenv
openai
faiss-cpu
